Create 'orders' app and its respective models
The models are still in the state of a "vague idea".  The main models
for this app include:

  Orders        To manage details of a particular order placed by a
                customer at a particular outlet.

  OrderItems    To manage the items associated with a particular order.

Multiple foreign key dependencies for the models are missing and hence
the models need to be further solidified with the creation of the
dependent apps and their respective models, before we can proceed with
the migration and testing.